# feat(cubestore): Rockstore - optimize index scanning